+++ src/file_storage.cpp
@@ -127,7 +127,7 @@ namespace libtorrent
 			// split the string into the leaf filename
 			// and the branch path
 			branch_path = path.c_str();
-			branch_len = leaf - path.c_str();
+			branch_len = int(leaf - path.c_str());
 
 			// trim trailing slashes
 			if (branch_len > 0 && branch_path[branch_len - 1] == TORRENT_SEPARATOR)
@@ -176,7 +176,7 @@ namespace libtorrent
 		else
 		{
 			// yes we do. use it
-			e.path_index = p.base() - m_paths.begin() - 1;
+			e.path_index = int(p.base() - m_paths.begin() - 1);
 		}
 		if (set_name) e.set_name(leaf);
 	}
@@ -415,7 +415,7 @@ namespace libtorrent
 
 		TORRENT_ASSERT(file_iter != m_files.begin());
 		--file_iter;
-		return file_iter - m_files.begin();
+		return int(file_iter - m_files.begin());
 	}
 
 	char const* file_storage::file_name_ptr(int index) const
@@ -446,7 +446,7 @@ namespace libtorrent
 
 		// in case the size is past the end, fix it up
 		if (std::int64_t(target.offset + size) > m_total_size)
-			size = m_total_size - target.offset;
+			size = int(m_total_size - target.offset);
 
 		auto file_iter = std::upper_bound(
 			m_files.begin(), m_files.end(), target, compare_file_offset);
@@ -461,7 +461,7 @@ namespace libtorrent
 			if (file_offset < std::int64_t(file_iter->size))
 			{
 				file_slice f;
-				f.file_index = file_iter - m_files.begin();
+				f.file_index = int(file_iter - m_files.begin());
 				f.offset = file_offset
 #ifndef TORRENT_NO_DEPRECATE
 					+ file_base_deprecated(f.file_index)
@@ -854,7 +854,7 @@ namespace libtorrent
 
 	sha1_hash file_storage::hash(internal_file_entry const& fe) const
 	{
-		int index = &fe - &m_files[0];
+		int index = int(&fe - &m_files[0]);
 		if (index >= int(m_file_hashes.size())) return sha1_hash(nullptr);
 		return sha1_hash(m_file_hashes[index]);
 	}
@@ -867,21 +867,21 @@ namespace libtorrent
 
 	time_t file_storage::mtime(internal_file_entry const& fe) const
 	{
-		int index = &fe - &m_files[0];
+		int index = int(&fe - &m_files[0]);
 		if (index >= int(m_mtime.size())) return 0;
 		return m_mtime[index];
 	}
 
 	int file_storage::file_index(internal_file_entry const& fe) const
 	{
-		int index = &fe - &m_files[0];
+		int index = int(&fe - &m_files[0]);
 		TORRENT_ASSERT_PRECOND(index >= 0 && index < int(m_files.size()));
 		return index;
 	}
 
 	void file_storage::set_file_base(internal_file_entry const& fe, std::int64_t off)
 	{
-		int index = &fe - &m_files[0];
+		int index = int(&fe - &m_files[0]);
 		TORRENT_ASSERT_PRECOND(index >= 0 && index < int(m_files.size()));
 		if (int(m_file_base.size()) <= index) m_file_base.resize(index + 1, 0);
 		m_file_base[index] = off;
@@ -889,7 +889,7 @@ namespace libtorrent
 
 	std::int64_t file_storage::file_base(internal_file_entry const& fe) const
 	{
-		int index = &fe - &m_files[0];
+		int index = int(&fe - &m_files[0]);
 		if (index >= int(m_file_base.size())) return 0;
 		return m_file_base[index];
 	}
@@ -897,7 +897,7 @@ namespace libtorrent
 	std::string file_storage::file_path(internal_file_entry const& fe
 		, std::string const& save_path) const
 	{
-		int const index = &fe - &m_files[0];
+		int const index = int(&fe - &m_files[0]);
 		return file_path(index, save_path);
 	}
 
@@ -922,7 +922,7 @@ namespace libtorrent
 	}
 
 	file_entry file_storage::at(file_storage::iterator i) const
-	{ return at_deprecated(i - m_files.begin()); }
+	{ return at_deprecated(int(i - m_files.begin())); }
 #endif // TORRENT_NO_DEPRECATE
 
 	void file_storage::reorder_file(int const index, int const dst)
@@ -990,8 +990,8 @@ namespace libtorrent
 
 				if (best_match != i)
 				{
-					int const index = best_match - m_files.begin();
-					int const cur_index = i - m_files.begin();
+					int const index = int(best_match - m_files.begin());
+					int const cur_index = int(i - m_files.begin());
 					reorder_file(index, cur_index);
 					i = m_files.begin() + cur_index;
 				}
@@ -1028,8 +1028,8 @@ namespace libtorrent
 						// look for files <= pad_size, which never is greater than
 						// alignment
 						TORRENT_ASSERT(best_match != i);
-						int index = best_match - m_files.begin();
-						int cur_index = i - m_files.begin();
+						int index = int(best_match - m_files.begin());
+						int cur_index = int(i - m_files.begin());
 						reorder_file(index, cur_index);
 						i = m_files.begin() + cur_index;
 						i->offset = off;
@@ -1081,7 +1081,7 @@ namespace libtorrent
 		, std::int64_t& offset
 		, int& pad_file_counter)
 	{
-		int const cur_index = i - m_files.begin();
+		int const cur_index = int(i - m_files.begin());
 		int const index = int(m_files.size());
 		m_files.push_back(internal_file_entry());
 		++m_num_files;
@@ -1135,7 +1135,7 @@ namespace libtorrent
 		// contained within the last file.
 		int const end_piece = (file == fs.num_files() - 1)
 			? fs.num_pieces()
-			: (range.piece * piece_size + range.start + file_size + 1) / piece_size;
+			: int((range.piece * piece_size + range.start + file_size + 1) / piece_size);
 		return std::make_tuple(begin_piece, end_piece);
 	}
